Initial Commit of kicad2itead 2019 update
This was a script created four years ago to quickly convert a directory of Gerber output files from KiCAD over to iTead's (at https://itead.cc) naming conventions and compress them. Also, it copies the files to a directory named "backup/" in the current directory if possible before renaming the files.